# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Public REST API for Legend of Elements Wiki. Unauthenticated GET methods return encyclopedia
articles, categories, search hits, stats, and tool links. Write methods require an admin session cookie.

 * OpenAPI spec version: 0.1.0
 */
export type ProblemError = {
  /** Machine-readable error code */
  code: string;
  /** Human-readable explanation */
  message: string;
  /** How an agent should recover */
  hint: string;
};

export interface Problem {
  error: ProblemError;
}

export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  title: string;
  slug: string;
  category: string;
  content: string;
  excerpt: string;
  tags: string[];
  featured: boolean;
  viewCount: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  content: string;
  excerpt: string;
  tags: string[];
  featured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ArticleUpdate {
  title?: string;
  slug?: string;
  category?: string;
  content?: string;
  excerpt?: string;
  tags?: string[];
  featured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  icon: string;
  articleCount: number;
}

export interface WikiStats {
  totalArticles: number;
  totalCategories: number;
  totalViews: number;
  recentEdits: number;
}

export interface Tool {
  id: number;
  name: string;
  description?: string;
  href: string;
  external: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface ToolInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  href: string;
  external?: boolean;
  sortOrder?: number;
}

export interface ToolUpdate {
  name?: string;
  description?: string;
  href?: string;
  external?: boolean;
  sortOrder?: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Problem;

/**
 * Admin session required
 */
export type UnauthorizedResponse = Problem;

/**
 * Resource not found
 */
export type NotFoundResponse = Problem;

export type ListArticlesParams = {
/**
 * Category slug such as cultivation or classes
 */
category?: string;
/**
 * Tag that must be present on the article
 */
tag?: string;
/**
 * Maximum number of articles to return
 * @minimum 1
 * @maximum 200
 */
limit?: number;
/**
 * Number of articles to skip
 * @minimum 0
 */
offset?: number;
};

export type ListRecentArticlesParams = {
/**
 * Maximum number of articles to return
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type SearchArticlesParams = {
/**
 * Search query string
 * @minLength 1
 */
q: string;
/**
 * Optional category slug to restrict results
 */
category?: string;
};