# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Legend of Elements Wiki API
  version: 0.1.0
  description: |
    Public REST API for Legend of Elements Wiki. Unauthenticated GET methods return encyclopedia
    articles, categories, search hits, stats, and tool links. Write methods require an admin session cookie.
  contact:
    name: Legend of Elements Wiki
    url: https://wiki-legendsofelements.vercel.app/contact
  license:
    name: Community / unofficial fan project
    url: https://wiki-legendsofelements.vercel.app/about
servers:
  - url: https://wiki-legendsofelements.vercel.app/api
    description: Production
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: articles
    description: Wiki article operations
  - name: categories
    description: Category operations
  - name: search
    description: Search operations
  - name: stats
    description: Wiki stats
  - name: tools
    description: Adjacent tools and sites
components:
  securitySchemes:
    adminCookie:
      type: apiKey
      in: cookie
      name: wiki_admin_session
      description: Signed admin session cookie set by POST /admin/login
  schemas:
    Problem:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, hint]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable explanation
            hint:
              type: string
              description: How an agent should recover
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Article:
      type: object
      required: [id, title, slug, category, content, excerpt, tags, featured, viewCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        content:
          type: string
        excerpt:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        viewCount:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ArticleInput:
      type: object
      required: [title, slug, category, content, excerpt, tags]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        content:
          type: string
        excerpt:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        content:
          type: string
        excerpt:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
    Category:
      type: object
      required: [id, name, slug, description, icon, articleCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        articleCount:
          type: integer
    WikiStats:
      type: object
      required: [totalArticles, totalCategories, totalViews, recentEdits]
      properties:
        totalArticles:
          type: integer
        totalCategories:
          type: integer
        totalViews:
          type: integer
        recentEdits:
          type: integer
    Tool:
      type: object
      required: [id, name, href, external, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        href:
          type: string
        external:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ToolInput:
      type: object
      required: [name, href]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        href:
          type: string
          minLength: 1
        external:
          type: boolean
        sortOrder:
          type: integer
    ToolUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        href:
          type: string
        external:
          type: boolean
        sortOrder:
          type: integer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Admin session required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns a JSON health payload when the API process is up. No authentication.
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      description: List encyclopedia articles, optionally filtered by category or tag. Public.
      parameters:
        - name: category
          in: query
          description: Category slug such as cultivation or classes
          schema:
            type: string
        - name: tag
          in: query
          description: Tag that must be present on the article
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of articles to return
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of articles to skip
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article
      description: Create an article. Requires admin session cookie.
      security:
        - adminCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /articles/featured:
    get:
      operationId: listFeaturedArticles
      tags: [articles]
      summary: Get featured/pinned articles
      description: Return articles marked featured. Public.
      responses:
        "200":
          description: Featured articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
  /articles/recent:
    get:
      operationId: listRecentArticles
      tags: [articles]
      summary: Get recently updated articles
      description: Return the most recently updated articles. Public.
      parameters:
        - name: limit
          in: query
          description: Maximum number of articles to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Recent articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by slug
      description: Fetch a single article by its unique slug and increment viewCount. Public.
      parameters:
        - name: slug
          in: path
          required: true
          description: Unique article slug
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article
      description: Partial update of an article. Requires admin session cookie.
      security:
        - adminCookie: []
      parameters:
        - name: slug
          in: path
          required: true
          description: Unique article slug
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      description: Permanently delete an article by slug. Requires admin session cookie.
      security:
        - adminCookie: []
      parameters:
        - name: slug
          in: path
          required: true
          description: Unique article slug
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories with article counts
      description: Return wiki categories and how many articles each contains. Public.
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /search:
    get:
      operationId: searchArticles
      tags: [search]
      summary: Search articles by title or content
      description: Full-text-ish search over title, excerpt, and content. Public. Requires q.
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
        - name: category
          in: query
          description: Optional category slug to restrict results
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
  /stats:
    get:
      operationId: getWikiStats
      tags: [stats]
      summary: Get wiki statistics
      description: Totals for articles, categories, views, and recent edits. Public.
      responses:
        "200":
          description: Wiki stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WikiStats"
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List adjacent tools and sites
      description: Community tools and related links shown on the wiki. Public.
      responses:
        "200":
          description: Tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
    post:
      operationId: createTool
      tags: [tools]
      summary: Create a tool link
      description: Add a tool link. Requires admin session cookie.
      security:
        - adminCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tools/{id}:
    put:
      operationId: updateTool
      tags: [tools]
      summary: Update a tool link
      description: Update a tool link by numeric id. Requires admin session cookie.
      security:
        - adminCookie: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric tool id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTool
      tags: [tools]
      summary: Delete a tool link
      description: Delete a tool link by numeric id. Requires admin session cookie.
      security:
        - adminCookie: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric tool id
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"