# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Public REST API for Legend of Elements Wiki. Unauthenticated GET methods return encyclopedia
articles, categories, search hits, stats, and tool links. Write methods require an admin session cookie.

 * OpenAPI spec version: 0.1.0
 */
import type { ProblemError } from './problemError';

export interface Problem {
  error: ProblemError;
}